pyproject.toml
[tool.poetry]
name = "stroke-prediction"
version = "0.1.0"
description = "Classify stroke patients"
authors = ["Joanna Khek <joannakhek@gmail.com>"]
readme = "README.md"
packages = [{include = "stroke_prediction", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.10"
pandas = ">=1.3.5"
kaggle = "^1.5.16"
pyyaml = "^6.0.1"
numpy = "^1.21.4"
scikit-learn = "^1.3.2"
loguru = "^0.7.2"
pydantic-settings = "^2.1.0"
matplotlib = "^3.8.2"
seaborn = "^0.13.1"
feature-engine = "^1.6.2"
xgboost = "^2.0.3"
hyperopt = "^0.2.7"
mlflow = "^2.9.2"
pytest = "^7.4.4"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.25.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"